RPP: Add tests for networkResourceCategory
Also, if a network request had Other as resourceType and a CSS/JS mimeType, it was mis-classified. This is likely impossible to hit in real traces, but I hit it in a synthetic rum-trace (where resourceType isn't available).

Change-Id: I5421ebc73bee89df12f2ec679c6ce9c3d7d16daf
Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6011774
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Paul Irish <paulirish@chromium.org>