Add update scheduling support to widgets
This CL adds support for RenderCoordinator to widgets. This is done by adding a new `update()` method to the `Widget` class that schedules a write job with the RenderCoordinator. The `update()` method returns a promise that resolves when the update is complete.

This CL also updates all of the legacy widgets that extend the `Widget` class to use the new `update()` method. This ensures that all of the legacy widgets are using the RenderCoordinator to schedule their updates.

This CL is part of a larger effort to improve the performance of the DevTools. By using the RenderCoordinator, we can ensure that the legacy widgets are only updated when they are visible on the screen. This can significantly improve the performance of the DevTools, especially for large and complex pages.

Bug: 301364727
Change-Id: I3b638aadf169a37a0328cfe16e3c462a709bc0da
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6021958
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>