[RPP] Change shortcuts list when a different navigation is selected
At the moment it is not obvious which shortcuts are for which action and some shortcuts are not displayed correctly. I will fix the css and a a test in a follow up.
Screencast: http://screencast/cast/NTAwODM2MTE5MDg1MDU2MHxkZmZmOGJlOS1hOQ


Bug: 313757601
Change-Id: Ibdc6c124908c2d8b5f0dc4af15341a3b868a32e9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6038621
Reviewed-by: Nancy Li <nancyly@chromium.org>
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>