[AI Assistance] Add streaming to answers and update auto scroll
Streaming changes:
1 - Updated `aidaFetch` to be an AsyncGenerator that
yields the partial responses after parsing them.
2 - We start yielding partial `ANSWER` responses from
`aidaFetch` which is then rendered on the UI.
3 - Updated `FreestylerPanel`s `run` to set `isLoading`
to `true` on `#doConversation` unconditionally and
set it `false` after the generation loop finishes.
4 - Added `finishAnimations` method to `MarkdownView`
so that users don't need to wait for animation to
finish after the answer is here.

Scroll changes:
1 - Updated scroll behavior in `#doConversation` method to
scroll to bottom automatically when there is a new
query submitted or when the side effect dialog is shown.
2 - Updated auto scroll behavior to pin to bottom as long
as user didn't scroll manually to top.
3 - Implemented pinning to bottom behavior by adding
a resize observer to message container (since with
animation, its height grows automatically and we want to
keep the scroll pinned to bottom as long as user didn't
scroll manually to top).

Considered using `overflow-anchor` for the pinning behavior,
however, it doesn't handle the case when the
scroller is just shown so we needed JS to trigger that
since we'll need JS either case, I decided
to go with a JS only solution for more control.

Fixed: 375556018, 373832629, 367651110
Change-Id: Ie93d32e03eeec93868bb8afcbe27f5c56b6c686d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6017888
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>