i18n: fall back to en-US for simple strings
This CL fixes an edge case where a UIString has had a placeholder
removed. On Canary (where we do not have up to date translations in all
locales), this caused an error.

With this CL we will now catch this error and fall back to en-US.

Bug: none
Change-Id: Idfb05255227c3006dd5dd4b5c200073864b7bf25
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6054628
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>