RPP: add AnimationFramesHandler
This CL adds the first pass at AnimationFrames parsing.

It pairs AnimationFrame begin & end events based on the fact that they
are a stack. It also does this per PID+TID to prevent against merging
incorrectly from different threads.

Additionally, it uses the ID field that is on the begin event + the
AnimationFrame::Presentation event to pair those up. This allows us to
know the time that a given frame was shown to the user.

Bug: 352242870
Change-Id: Id8a120d5c7b13cd4751e23cbccc83ccee8d4c265
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6054290
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>