[RPP] Add AsyncCallStacksHandler
The handler will generalize async JS calls in the perf panel. Its output
will be the source for the InitiatorHandler's JS scheduling
relationships, so that they don't need to be hardcoded as they are
currently.

The output of this handler will also be the source for async call
stacks that will be shown in the details of JS entries (ProfileCalls)
but this require some reworking on it, which will happen in follow up
CLs.

Bug: 381391955
Change-Id: If1b3a845dccf35bbfd6ae2aa06e7baf4a282d5e0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6055479
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>