Fix command menu when used with an IME
The Enter key is used both to confirm a selection in an IME
but also to confirm a selection in the command menu. We need
to distinguish this to not confirm both selections on the
same Enter keyboard event.

Fixed: 370332848
Change-Id: Ic0b09f8bc3b334e74c6049050edf7981776694d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6054256
Auto-Submit: Yang Guo <yangguo@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>