[RPP] Track async scheduler for async JS calls
The AsyncCallStacksHandler now tracks the JS call that scheduled a
particular JS function. Basically, this adds the opposite direction of
the relationships we were tracking: before we had scheduler -> schedulee
this CL tracks schedulee -> scheduler.

This is done with the twist that instead of starting from a JS entry
point trace event (e.g. FunctionCall, EvaluateScript, etc.) we start
from the profile calls themselves.

This will be useful in reconstructing the full (async) callstack of a
given JS call in its details, which will be done in a follow up

Bug: 381392952
Change-Id: I4a3983e2cc21f8c3ccf3c856fade99ed845f3421
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6059857
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>