[browser_tests] Rethrow in setupLegacyFilesForTest
This is part of a 3-way change to get rid of DOMMessageQueue in
LoadLegacyFilesInFrontend [1].

"ExecJS" can await promises, so all we need is to make sure that the
async function promise is rejected when the imports fail. Once this
CL has landed we can remove the DOMMessageQueue and just call
"ExecJS" with the default "await promise" option.

The last step is to remove the try-catch alltogether in the frontend.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/devtools/devtools_browsertest.cc;l=221-231;drc=f633fa3cf093bc8b9eb3243e7b05ff0d64e4bff9

R=alexrudenko@chromium.org

Bug: 382003727
Change-Id: I7cee60082f285cf3fe0ea562424269da74f5b989
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6063120
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>