Improve spacing within SelectMenu
The motivation for this is to align the arrows for the two select menus
found in the Performance panel (cpu/network throttling). The menus are
the same width, but due to its content and the choice of
`justify-content` their arrows did not align.

https://imgur.com/a/jgX1bg3

Bug: None
Change-Id: I1cf7f08c66c99188adc20f984292ce763f628cf8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6069006
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>