[Freestyler] Do not suggest `setElementStyles` in the response
I have run the evals and did not see a change in its tendency
to run code with `setElementStyles`.

Tested the instruction with asking:
"I remember there's a command in DevTools to
change element styles directly. It starts
with 'setElement'. What's the full command
and how do I use it to change the font size
to 20px and the color to blue for the
selected element?"

Observed that now the LLM understands that this is an internal
function and answers appropriately.

Previously, it was saying "You can use `setElementStyles` like
this..."

Fixed: 373612171
Change-Id: I7cc9a3406e9eef3fc18a9276d4e831c6098ad1fd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6070367
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>