[Elements] Rename `ComputedStyleChanged` event to `CSSModelChanged`
`ComputedStyleChanged` event does not directly reflect
the occasions where this event is emitted.

For example, it might be that a new StyleSheet added
to the CSSModel that didn't change the computed
styles of the node. So, we rename the event to be
`CSSModelChanged`.

Bug: 349566291
Change-Id: I341713ed591e0d88c1842ed5f6bb3cc0c32fae41
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6069996
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>