[Elements] Convert StylesSidebarPane to a regular class
ElementsPanel already is a singleton and we can provide
access to the StylesSidebarPanel it rendered through it.

This way, we can control the lifecycle of the StylesSidebarPane
directly from the ElementsPanel and do not need to
have a singleton class for it.

Bug: 349566291
Change-Id: Ie1ee75677a305712d482812ad910bc38c583afd9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6069998
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>