Extract async task name in AsyncJSCallsHandler
This was added to the new event [1] and will allow us to determine the
name of the asynchronous task. This is useful when displaying async
stacks as the name of the task is used as the connection between the
linked stack traces. See screenshot of the sources panel usage of it:
https://screenshot.googleplex.com/3Hc3apDfM6roR4D

[1] https://chromium-review.googlesource.com/c/v8/v8/+/6054982

Bug: 381392952
Change-Id: I4e037a7f121868e5b3c077ea9ba656b18d63d6ed
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6070373
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>