[RPP] Introduce StackTraceForEvent helper
It extracts the full call stack (including async calls) of a given event
(at the moment only JS calls / ProfileCalls). We will use this to
helper to display the call stack of a given event in its details. In the
future support for extension entries will be added to it.

Bug: 381392952
Change-Id: I1410b2ed0f85bb8f9e4799ae7b0606bbc7f247f5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6074770
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>