RPP: fix CLS Culprits Insight not rendering
This CL fixes a bug in the CLS Culprits insight where it tried to read
`this.bounds` which was undefined. I think TS expects the `get` to exist
when the `set` does (we hit this a lot in our components). To fix this I
added `get bounds`.

I then added a unit test to verify the fix and to ensure it renders. I
then went down a bit of a rabbit-hole of styling; the <li> elements for
the culprits were not wrapped in a <ul> so I added that and tidied up
some of the margins.

Bug: none
Change-Id: I38dad120760f2f8bc13c63c31f07dcafe11703eb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6074609
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>