[RPP] Ensure entry re-highlights after clearing selection
If the user clicks an entry with an ENTRY_OUTLINE overlay, it will be
replaced with a ENTRY_SELECTED overlay. However, if the selection is
cleared (e.g. clicking somewhere in blank space) the ENTRY_OUTLINE will
not immediately be restored.

This CL resolves the issue by keeping the original ENTRY_OUTLINE
rendered, and using CSS to ensure that the ENTRY_SELECTED overlay is
always rendered on top of it.

Bug: None
Change-Id: Ia5263089b0e00433e2c500643646bcff019fd5ef
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6084661
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>