Fix network category filters when using locales other than en-US
The CL https://crrev.com/c/5938364 changed resource categories to
use an un-localized name as a key, as this was messsing with VE
logging.

This broke the filter UI for non-english locales, as the
NetworkLogView would retrieve the "title" instead of the "name" from
each NetworkRequest to check if a filter applies.

This CL fixes this and adds a regression test.

R=dsv@chromium.org

Fixed: 382596366
Change-Id: I039614285a99b948ea21b224255be557ad5a5fb4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6085916
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>