[RPP] Parse trace event for consoleTask.run
Recently, an event was added as an entrypoint of the task being run
with the console.createTask / run API [1]. This allows us to simplify
our parsing of async js tasks, because now all of these have a single
trace event entry point for the scheduled task. Before, for consoleTask.run we had to assign all the top "run" profile calls as the scheduled task, because of limitations with sampling it's possible that multiple profile calls are made from a single call to the API. With tracing we don't have this limitation, yielding a more simple code.

[1] https://chromium-review.googlesource.com/c/v8/v8/+/6054136

Bug: 381392952
Change-Id: I18178c17657dc0e6c2733729f4692f214f5654c5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6084415
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>