[RPP] Fix Entries Link not updating position in the Network Track
The number of levels in the network track changes and it caused the annotated entries to be considered not existant in the method updating the annotation position. It caused the positioning method to exit early.
Instead, check if the entry is visible before quering the entry parameters and if it is not visible, draw the arrow to the coordinate returned by `yCoordinateForNotVisibleEntry`.

Bug: 378174655
Change-Id: I5575a4b343fcec275fce68f0d55371ed66e7c891
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6079113
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>