[RPP] disregard ignorelist in StackTraceForEvent util
The util received a call back to determine if a call frame is
ignorelisted, allowing the UI to pass ignorelists to the util before
rendering. However, this is not necessary because the JSPresentationUtil
we use already checks for ignorelists before creating the stack trace
element. Thus we can remove this ignorelist logic from the
StackTraceForEvent util and its usages.

Bug: 374732385
Change-Id: I9d7653a7de25e71dc8a53feb7825a0496c62c8b9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6091370
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>