[RPP] Remove ExtensionDataGatherer
It serves as an intermediary layer between the parsedTrace data and the
UI, and it adds some overhead to be properly managed because of our lack
of proper support for this type of utils in the RPP architecture.

The reason it exists at all is in preparation to support extending the
RPP with Chrome extensions, as that was the inital plan. But at the
moment, we are focusing on extending the panel with web APIs
(performance.measure, console.timeStamp, etc.), which makes the data available via trace events, so right now and for the foreseable future, this extra layer doesn't bring any value to outweight its maintenance cost. Thus this CL.

Bug: 381861194
Change-Id: I24a1bc3f7405dc5c8155e7f554007fa48c33583c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6088062
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>