[RPP] Show full stack traces for extension events
The StackTraceForEvent helper now supports extension entries, and its
output is used in their details. Just like with profile calls, we want
the full stack trace, including async calls. For this reason we don't
use only the stack trace available in the trace event, since that
contains only the synchronous function. Frameworks and third parties in
general use async tasks schedulers, meaning that to improve attribution
to authored source code in extension entries support for async call
stacks is needed.

To build the async call stack of an extension entry, we look for the JS
function call (ProfileCall) where the call to the API was made (aka the
call point to performance.mark/measure), then take the callstack of this
profile call as the call stack of the extension entry. The assignment of
the profile call to the extension entry is done based on the (sync)
stack trace available in the extension entry trace event and on the time
the API was called, this data is used to look for the closest matching
profile call.

Bug: 381861194
Change-Id: Ia3cfa9d12ecc6d6d7fcabb9faaa0e0f39613dab6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6084695
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>