RPP: fix lcp image request <> navigation relation
The below bug was reported where the LCP Insight was using the wrong
image request for the LCP event. This was because the
LargestImagePaintHandler tied events to the current navigation, but if
the page being loaded has multiple frames, the "current navigation"
isn't that helpful and instead we need to track "current navigation per
frame" in order to not have data from frame A get mixed up with data
from frame B.

The bug below is caused by the "current navigation" becoming the loading
of an iframe, which happens before the actual LCP image is loaded. That
means we mistakenly relate the actual LCP image to an entirely different
and unrelated navigation.

The fix here is to track the current navigation per frame. Then when we
look up and relate the LCP Image request to the navigation, we aren't
mistakenly mixing requests across different frames.

I also landed an example in the stories repo
https://github.com/ChromeDevTools/performance-stories/pull/38 to give us
a test case for this and used it in the unit test.

Fixed: 384000716
Change-Id: Ica187428545ec1d783ea3a050d0e837ad6f3f729
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6091385
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>