[RPP] Set up 3p cache
This starts building the caches from the NetworkHandler down to the
RendererHandler. Also moves URLForEntry so that it could be depended on
by the Handlers.

This is not yet used anywhere, but this sets the data that we need
to be referenced in follow up features like the checkbox, summary
details table, etc.

This also doesn't yet handle keeping the cache up to date via
sourceMaps, this will be worked through on a separate cl

Bug:381960419
Change-Id: I9629f5a511908166ad104aad37b2590c04b0a23f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6069068
Reviewed-by: Adam Raine <asraine@chromium.org>
Commit-Queue: Adriana Ixba <aixba@chromium.org>