[RPP] Add manual parsing of JS scheduling events for backwards compat
A partial revert of [1] to ensure async calls are computed in old
traces. We'll keep this backwards compatibility behavior for some time
and then remove it since traces taken with chrome 133 will include the
new events that allow us to generalize the computation of async calls.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6054289

Bug: 383974422
Change-Id: I6844dfe0d4ffb9d65b23689e09e994e1b81ffb7b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6091377
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>