[scopes] Add 'findOriginalFunctionName' method
This CL adds a new method that takes a generated position and spits out
the original function name of the surrounding function.

The method works with both GeneratedRanges and mappings. The
reason for this is twofold: "pasta" source map scopes only have
"OriginalScopes" and no generated ranges. And the proposed extension
API in https://crrev.com/c/5774044 also does not contain generated
ranges.

In both of these cases the implementation utilizes the normal source
map "mappings" to get to the original function scope.

As a follow-up, we'll consolidate the "ScopesEntry" interface (for
"pasta" source maps) and the "SourceMapScopesInfo" class (for the
official "scopes" proposal). We'll do this by converting the "pasta"
scopes into "OriginalScope" trees.

Bug: 40277685
Change-Id: I1a50eaca5c719616db3c2f7eb4e0227cddee2e2f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6108439
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>