Disable device mode for chrome:// pages
chrome:// pages do not target mobile devices, yet have been a
frequent source of bug reports for device mode.

When the user navigates to a chrome:// page, device mode is turned off
(if on) and disabled. If device mode was previously (in the current
session) on when navigating to a chrome:// page, it will be turned on
again when navigating to a non-chrome:// page.

Fixed: 40186276
Change-Id: I7d9ddb2874b91ca04813d5135eb1da8f495a2e16
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6110495
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Auto-Submit: Yang Guo <yangguo@chromium.org>