[GM3Restyling] Update tab bars
This is not yet moving the icons to the right, as intended in the designs. This is because for the settings and the vertical layout, we need the icons on the left. After this CL we can refactor it two allow icons on both sides.

Before: https://imgur.com/9lf6UGZ
After: https://imgur.com/a/YEHoNT9
Bug: 325442723
Change-Id: Ie2bd99726e767598258d3c8d34d648ba793ca781
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6089109
Reviewed-by: Nancy Li <nancyly@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Marcel Pütz <marcelpuetz@chromium.org>