[Flame chart] Fix the bug for memory panel hovering
The memory panel will not have a group, but just append the entries, So the |HoverType| will always be |OUTSIDE_TRACKS|.
But each entry can still be hovered/clicked when users are not hovering any track.
So the fix is still do the hovering/clicking behavior for entries even no track is hovered.

Fixed: 360795014
Change-Id: Ia0696061d99b5b1e2c90f9e3b237ac6ae430bdf0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6110498
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>