[issues] Prepare the front-end for `AffectedRequest` change.
We are going to change the `browser_protocol.pdl` in a follow-up
back-end CL to make the `url` field of `AffectedRequest` mandatory and
the `requestId` field optional. This will remove the confusion with
issues for early CORS errors that happen before the `ResourceLoader` is
created, and for which we simply don't have a request ID (and will never
have one).

With this CL, the front-end works independent of whether the `requestId`
is provided or not, and can even deal with absent `url` fields (which
will not happen anymore after the back-end CL is landed).

Drive-by-fix: Tweak the CSS a bit, to ensure that the `<button>`
inherits the correct text settings.

Bug: 385693302
Change-Id: Ib529fd081e582fdaa5bea40cd902cd182b30b9a2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6111100
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>