[cleanup] Remove `UI.UItils.registerCustomElement()`.
As part of the Vision for the Chrome DevTools UI Engineering[^1],
we need to migrate off of built-in custom elements. Which also has
the additional benefit that our UI will work in Safari (part of
this was already fixed with https://crrev.com/c/6028362 earlier).

This CL replaces the previous built-in custom elements with proper
autonomous custom elements: `BezierSwatch`,`DevToolsIconLabel`,
`DevToolsSmallBubble`, and `DevToolsCloseButton`.

The `HistoryInput` component only added special handling for keydown
and input events, and otherwise was just a simple `HTMLInputElement`.
This migrates the class to a simple helper function that does the
same (without requiring a built-in custom element, which we're
trying to migrate off of).

The `DevToolsSlider` and `DevToolsRadioButton` built-in custom
elements are removed in favor of simple helpers that just create
the appropriate `HTMLInputElement`s. The styles for these are
moved to the `inspectorCommon.css`. The custom elements were
really overkill for something trivial as these and only added
confusion. The component server and documentation now also describe
sliders and radio buttons.

This change is functionally neutral and only replaces implementation
details.

Drive-by-fix: Fix a bunch of outdated / incorrect CSS along the way.

[^1]: http://go/chrome-devtools:ui-engineering-proposal

Bug: 386453980
Change-Id: I7ba9f9c156e5f0f9396c0c6eb3d2febc8d371b33
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6114579
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>