Add left margin to <length> tool for easier selection
Currently it's a little bit hard to drag-select the property value of a
CSS declaration when there is a <length> tool in the value, because the
drag-select is overriden by the drag-to-change-value behavior when the
mouse hovers over the property value. This CL adds a little bit more
space before the <length> tool, so that the mouse is easier to drag-
select before the <length> tool.

Fixed: 356039091
Change-Id: I391293864d79c6399e2be1f0c16a3adde1b43f03
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6126695
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>