[ui] Handle update cycle (promise) within `Widget`.
This is a follow-up change to https://crrev.com/c/6021958, which moves
the update promise logic into the `Widget` class and no longer requires
a leaky abstraction in the `RenderCoordinator`.

As part of this, we also add a `LitElement`-inspired `updateCompleted`
property, that yields the latest update promise, instead of the previous
`pendingUpdate()` method.

Bug: 387186170
Change-Id: Ie84736240bad81a5e74a17ea1aeec2cf5e5da754
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6133000
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>