[cleanup] Expose `RenderCoordinator` methods as functions.
This is step one to making the `RenderCoordinator` more ergonomic to use
(and our codebase slightly more readable): Instead of going via the
singleton all the time, just expose the `RenderCoordinator` methods as
functions on the `RenderCoordinator` module. That makes code shorter and
easier on the eyes, and also doesn't leave the engineer wondering why on
earth the `RenderCoordinator` is a class.

The second step will be to replace the implementation of the render
coordinator with just functions instead of the singleton.

Bug: 387186170
Change-Id: I011bbbb5bae21f01762d15c9a8d0a7abff45cac2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6120416
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>