Support console timestamp event dispatched at V8::Console
console.timeStamp will be used as a fast API to extend the performance
panel [1]. Right now, the call dispatches two trace events, one in Blink
and one in V8. Before this CL the timings added to the timeline for
console.timeStamp calls were taken from the event dispatched from Blink.

However, we have recently modified the V8 event to include 5 extra
params in accordance to the proposed extension to the console.timeStamp
API in the linked doc [2]. We did it in the V8 event to avoid the
potential overhead of passing the params to Blink from V8 and keep the
API as fast as possible.

This CL replaces the usage of the console timestamp event in Blink for
the one in V8, and updates the event type accordingly. This is the first
step towards adding support for extending the Performance panel with the
console.timeStamp API.

Note: With this done we can also delete the trace event dispatched in
Blink, since it's no longer used.

[1] go/cpq:console-timestamp
[2] https://chromium-review.googlesource.com/c/v8/v8/+/6110720

Bug: 377884142
Change-Id: I5f2a0ebd212ae5169b5c3f88bdda6d0b64ae00eb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6141613
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>