[RPP] Apply optimizations to statsForTimeRange
This trace is of the performance panel: https://trace.cafe/t/zP4zA7sNLb.
One thing of interest is that `statsForTimeRange` takes 3.5s, which is
33% of the `loadingComplete` callback.

One thing that jumped out to me is that 33% of `statsForTimeRange` is
spent inside `eventTimingsMilliSeconds`. The work this method must do
is simple, but it calls another method in an effort to DRY the code.
This creates an extra object allocation. Additionally, it uses the
timing tag and conversion functions multiple times. Given these are
called 100,000s of times during the processing of this trace, that adds
up. Un-DRYing the code and should have a good effect.

Additionally, `statsForTimeRange` can get away with calling
`eventTimingsMicroseconds` instead of `eventTimingsMilliSeconds`, which
saves 3 divisions per trace event.

Bug: None
Change-Id: Ie7ace912ebaac6c879233a809223710264081ed0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6154635
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>