DevTools: Add tests for DataGridWithPreview component.
Add tests for the DataGridWithPreview component. A lot of the logic
for actually rendering the data is in DataGrid, so I plan to expand test
coverage there in a second CL. These changes focus on test coverage for
the preview behavior in the bottom split pane widget.

I identified an unset (and unused in all places but DataGridWithPreview)
property `key` on DataGridNode. With that in mind, I simply removed it
to avoid accidental usage in the future.

Bug: 377942776
Change-Id: Id0da39ce8b21a1737c030f95eeed8390c77ed9a1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5999722
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Oliver Dunk <oliverdunk@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>