Show network request for cookie report line
This is the same functionality that the cookie table has within the
network panel. For each entry in the cookie report, there will be a
right click context menu option that will show the user the network
requests that contain cookies with the same cookie name AND cookie
domain.

Bug:b/365737493
Change-Id: I445e0b253f8ffd4f25d2bcc9a6560d4ab1ba39a8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6140936
Reviewed-by: Shuran Huang <shuuran@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>