[RPP code health] Remove the unnecessary map assignment
The spread syntax (...) will make a copy of the original object, which will use too much memory in our case because we have a big number of entries.
Originally we use spread syntax because we want to combine two maps. So the improvement is to check both map when we use the combined map.

Bug: 383702926
Change-Id: I191471ce8b7b33d978c61ea335acdac5fb6d412b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6152579
Commit-Queue: Nancy Li <nancyly@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>