[M133][WebVitals] Do not store interaction target nodes
This is a manual modification of the web-vitals.js library to prevent
it from retaining DOM nodes. In some cases the DOM node may be removed
before DevTools can make a reference for it, but that is preferable to
keeping the node in memory when the user does not expect it.

Bug: 376777343, 389082843
Change-Id: I507273c93a4bbe5f2f5111e1a17e179efc583ef4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6159204
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
(cherry picked from commit 82e6b18a383ab9f60187ab54a3b84b21a4dce714)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6166303
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>